requirements.txt
Click
requests
python-dateutil